DRY up common corporate information page type methods
Corporate information pages and worldwide organisation pages both have a
corporate information type. This moves any shared methods into a module.
